+++ react/components/CostCenterDetails.tsx
@@ -1,6 +1,6 @@
 import type { FunctionComponent } from 'react'
 import React, { useEffect, useState, useContext } from 'react'
-import { useQuery, useMutation } from 'react-apollo'
+import { useQuery, useMutation, useLazyQuery } from 'react-apollo'
 import {
   Layout,
   PageHeader,
@@ -24,12 +24,15 @@ import EditAddressModal from './EditAddressModal'
 import DeleteAddressModal from './DeleteAddressModal'
 import DeleteCostCenterModal from './DeleteCostCenterModal'
 import GET_COST_CENTER from '../graphql/getCostCenterStorefront.graphql'
+import GET_ORGANIZATION from '../graphql/getOrganizationStorefront.graphql'
 import UPDATE_COST_CENTER from '../graphql/updateCostCenter.graphql'
 import DELETE_COST_CENTER from '../graphql/deleteCostCenter.graphql'
+import GET_PERMISSIONS from '../graphql/getPermissions.graphql'
 
 const localStore = storageFactory(() => localStorage)
 let isAuthenticated =
-  JSON.parse(String(localStore.getItem('orderquote_isAuthenticated'))) ?? false
+  JSON.parse(String(localStore.getItem('b2b-organizations_isAuthenticated'))) ??
+  false
 
 const storePrefix = 'store/b2b-organizations.'
 
@@ -78,7 +81,7 @@ const CostCenterDetails: FunctionComponent = () => {
       sessionResponse?.namespaces?.profile?.isAuthenticated?.value === 'true'
 
     localStore.setItem(
-      'orderquote_isAuthenticated',
+      'b2b-organizations_isAuthenticated',
       JSON.stringify(isAuthenticated)
     )
   }
@@ -93,6 +96,7 @@ const CostCenterDetails: FunctionComponent = () => {
     showToast({ message: translatedMessage, duration: 5000, action })
   }
 
+  const [permissionsState, setPermissionsState] = useState([] as string[])
   const [loadingState, setLoadingState] = useState(false)
   const [costCenterName, setCostCenterName] = useState('')
   const [addresses, setAddresses] = useState([] as Address[])
@@ -120,15 +124,41 @@ const CostCenterDetails: FunctionComponent = () => {
     ssr: false,
   })
 
+  const [getOrganization, { data: organizationData }] = useLazyQuery(
+    GET_ORGANIZATION
+  )
+
+  const {
+    data: permissionsData,
+    // loading: permissionsLoading,
+  } = useQuery(GET_PERMISSIONS, { ssr: false })
+
   const [updateCostCenter] = useMutation(UPDATE_COST_CENTER)
   const [deleteCostCenter] = useMutation(DELETE_COST_CENTER)
 
   useEffect(() => {
-    if (addresses.length || !data?.getCostCenterById?.addresses?.length) return
+    if (
+      addresses.length ||
+      !data?.getCostCenterByIdStorefront?.addresses?.length
+    )
+      return
 
-    setCostCenterName(data.getCostCenterById.name)
-    setAddresses(data.getCostCenterById.addresses)
+    setCostCenterName(data.getCostCenterByIdStorefront.name)
+    setAddresses(data.getCostCenterByIdStorefront.addresses)
+    getOrganization({
+      variables: { id: data.getCostCenterByIdStorefront.organization },
+    })
   }, [data])
+
+  useEffect(() => {
+    if (!permissionsData) return
+
+    const { permissions = [] } = permissionsData.checkUserPermission ?? {}
+
+    if (permissions.length) {
+      setPermissionsState(permissions)
+    }
+  }, [permissionsData])
 
   const handleUpdateCostCenter = () => {
     setLoadingState(true)
@@ -159,7 +189,7 @@ const CostCenterDetails: FunctionComponent = () => {
       .then(() => {
         navigate({
           page: 'store.organization-details',
-          params: { id: data.getCostCenterById.organization },
+          params: { id: data.getCostCenterByIdStorefront.organization },
         })
       })
       .catch(error => {
@@ -328,11 +358,14 @@ const CostCenterDetails: FunctionComponent = () => {
       pageHeader={
         <PageHeader
           title={formatMessage(messages.pageTitle)}
-          linkLabel={formatMessage(messages.back)}
+          linkLabel={
+            organizationData?.getOrganizationByIdStorefront?.name ??
+            formatMessage(messages.back)
+          }
           onLinkClick={() => {
             navigate({
               page: 'store.organization-details',
-              params: { id: data.getCostCenterById.organization },
+              params: { id: data.getCostCenterByIdStorefront.organization },
             })
           }}
         >
@@ -340,7 +373,11 @@ const CostCenterDetails: FunctionComponent = () => {
             <Button
               variation="primary"
               isLoading={loadingState}
-              disabled={!costCenterName || !addresses.length}
+              disabled={
+                !costCenterName ||
+                !addresses.length ||
+                !permissionsState.includes('create-cost-center-organization')
+              }
               onClick={() => handleUpdateCostCenter()}
             >
               <FormattedMessage id="store/b2b-organizations.costCenter-details.button.save" />
@@ -350,6 +387,9 @@ const CostCenterDetails: FunctionComponent = () => {
             variation="danger"
             isLoading={loadingState}
             onClick={() => handleDeleteCostCenterModal()}
+            disabled={
+              !permissionsState.includes('create-cost-center-organization')
+            }
           >
             <FormattedMessage id="store/b2b-organizations.costCenter-details.button.delete" />
           </Button>
@@ -364,6 +404,9 @@ const CostCenterDetails: FunctionComponent = () => {
           onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
             setCostCenterName(e.target.value)
           }}
+          readOnly={
+            !permissionsState.includes('create-cost-center-organization')
+          }
           required
         />
       </PageBlock>
@@ -384,31 +427,37 @@ const CostCenterDetails: FunctionComponent = () => {
                       </AddressRules>
                     </div>
                     <div>
-                      <ActionMenu
-                        buttonProps={{
-                          variation: 'tertiary',
-                          icon: <IconOptionsDots color="currentColor" />,
-                        }}
-                        options={options(address.addressId)}
-                      />
+                      {permissionsState.includes(
+                        'create-cost-center-organization'
+                      ) && (
+                        <ActionMenu
+                          buttonProps={{
+                            variation: 'tertiary',
+                            icon: <IconOptionsDots color="currentColor" />,
+                          }}
+                          options={options(address.addressId)}
+                        />
+                      )}
                     </div>
                   </div>
                 </Card>
               </div>
             )
           })}
-          <div className="w-25 ma3">
-            <Card>
-              <div className="flex justify-center">
-                <Button
-                  variation="primary"
-                  onClick={() => handleNewAddressModal()}
-                >
-                  <FormattedMessage id="store/b2b-organizations.costCenter-details.address.new" />
-                </Button>
-              </div>
-            </Card>
-          </div>
+          {permissionsState.includes('create-cost-center-organization') && (
+            <div className="w-25 ma3">
+              <Card>
+                <div className="flex justify-center">
+                  <Button
+                    variation="primary"
+                    onClick={() => handleNewAddressModal()}
+                  >
+                    <FormattedMessage id="store/b2b-organizations.costCenter-details.address.new" />
+                  </Button>
+                </div>
+              </Card>
+            </div>
+          )}
         </div>
       </PageBlock>