+++ react/components/OrganizationUsersTable.tsx
@@ -13,6 +13,7 @@ import RemoveUserModal from './RemoveUserModal'
 
 interface Props {
   organizationId: string
+  permissions: string[]
 }
 
 interface CellRendererProps {
@@ -70,6 +71,7 @@ const messages = defineMessages({
 
 const OrganizationUsersTable: FunctionComponent<Props> = ({
   organizationId,
+  permissions,
 }) => {
   const { formatMessage } = useIntl()
   const { showToast } = useContext(ToastContext)
@@ -82,7 +84,9 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
   const [removeUserModalOpen, setRemoveUserModalOpen] = useState(false)
   const [usersState, setUsersState] = useState([])
 
-  const { data, refetch } = useQuery(GET_USERS, {
+  const canEdit = permissions.includes('add-users-organization')
+
+  const { data, loading, refetch } = useQuery(GET_USERS, {
     variables: { organizationId },
     ssr: false,
     skip: !organizationId,
@@ -189,9 +193,10 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
         fullWidth
         schema={getSchema()}
         items={usersState}
+        loading={loading}
         emptyStateLabel={formatMessage(messages.emptyState)}
         onRowClick={({ rowData }: CellRendererProps) => {
-          if (!rowData) return
+          if (!rowData || !canEdit) return
 
           setEditUserDetails({
             id: rowData.id,
@@ -210,6 +215,7 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
           newLine: {
             label: formatMessage(messages.new),
             handleCallback: () => setAddUserModalOpen(true),
+            disabled: !canEdit,
           },
         }}
       />